evaluation of optimization extracted
hdf5: compressed

+++ simnibs/simulation/fem.py
@@ -1191,17 +1191,24 @@ def _run_tms(mesh, cond, fn_coil, fields, matsimnibs, didt, fn_out, fn_geo,
     v.mesh = mesh
     out = calc_fields(v, fields, cond=cond, dadt=dAdt)
     mesh_io.write_msh(out, fn_out)
+    error = None
 
     # save hdf5 version to disk
     if write_hdf5:
-        fn_hdf5 = fn_out[:-3] + "hdf5"
-        if os.path.exists(fn_hdf5):
-            logger.warn(fn_hdf5 + " already exists. Removing file.")
-            os.remove(fn_hdf5)
-        logger.info("Writing .hdf5")
-        out.write_hdf5(fn_hdf5)
+        try:
+            fn_hdf5 = fn_out[:-3] + "hdf5"
+            if os.path.exists(fn_hdf5):
+                logger.warn(fn_hdf5 + " already exists. Removing file.")
+                os.remove(fn_hdf5)
+            logger.info("Writing .hdf5")
+            out.write_hdf5(fn_hdf5)
+        except:
+            logger.warn("Could not write {}.".format(fn_out))
+            error = 1
+
     if remove_msh:
-        os.remove(fn_out)
+        if not error:
+            os.remove(fn_out)
 
 
 def _finalize_global_solver():
